@@ -1,5 +1,5 @@
 // File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
+// ignore_for_file: type=lint
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
@@ -18,7 +18,7 @@ class DefaultFirebaseOptions {
   static FirebaseOptions get currentPlatform {
     if (kIsWeb) {
       throw UnsupportedError(
-        'DefaultFirebaseOptions have not been configured for macos - '
+        'DefaultFirebaseOptions have not been configured for web - '
         'you can reconfigure this by running the FlutterFire CLI again.',
       );
     }
@@ -51,7 +51,7 @@ class DefaultFirebaseOptions {
 
   static const FirebaseOptions android = FirebaseOptions(
     apiKey: 'AIzaSyB5Vitlm4m0HeeX2ZJp6IvhIG1xJlcPDS8',
-    appId: '1:752904600547:android:580353e0f00d6d70097ed6',
+    appId: '1:752904600547:android:790aa9982777dbc5097ed6',
     messagingSenderId: '752904600547',
     projectId: 'deliverytak',
     databaseURL: 'https://deliverytak-default-rtdb.firebaseio.com',
@@ -60,15 +60,13 @@ class DefaultFirebaseOptions {
 
   static const FirebaseOptions ios = FirebaseOptions(
     apiKey: 'AIzaSyCPN82WtFQHxvP-hIVA8ug9kK0mcOj27Wk',
-    appId: '1:752904600547:ios:fcad75da46ee73f0097ed6',
+    appId: '1:752904600547:ios:75e9c3a13c269542097ed6',
     messagingSenderId: '752904600547',
     projectId: 'deliverytak',
     databaseURL: 'https://deliverytak-default-rtdb.firebaseio.com',
     storageBucket: 'deliverytak.appspot.com',
-    androidClientId:
-        '752904600547-3lqg84h4ff8uudsu1ct61vlssubnmnpu.apps.googleusercontent.com',
-    iosClientId:
-        '752904600547-qn94i0n4r98jseghfmkmea9obrl9qfae.apps.googleusercontent.com',
-    iosBundleId: 'com.example.deliverytak',
+    androidClientId: '752904600547-3lqg84h4ff8uudsu1ct61vlssubnmnpu.apps.googleusercontent.com',
+    iosClientId: '752904600547-3t74a9oorbor8ifj66o2tns6k3vg15fa.apps.googleusercontent.com',
+    iosBundleId: 'com.arandas.deliverytak',
   );
 }
