@@ -1,15 +1,28 @@
 import '/exports.dart';
 
 class BrandModel {
-  String id, name;
+  String id, name, image;
 
-  BrandModel({required this.id, required this.name});
+  BrandModel({
+    required this.id,
+    required this.name,
+    required this.image,
+  });
 
-  factory BrandModel.fromJson(Map<String, dynamic> json) =>
-      BrandModel(id: json['id'], name: json['content']['title']);
+  factory BrandModel.fromDoc(DocumentSnapshot json) => BrandModel(
+        id: json.id,
+        name: json['name'],
+        image: json['image'],
+      );
 
-  factory BrandModel.fromDoc(DocumentSnapshot json) =>
-      BrandModel(id: json.id, name: json['content']['title']);
+  BrandModel copyWith({String? name, String? image}) => BrandModel(
+        id: id,
+        name: name ?? this.name,
+        image: image ?? this.image,
+      );
 
-  Map<String, dynamic> toJson() => {'id': id, 'name': name};
+  Map<String, dynamic> toJson() => {
+        'name': name,
+        'image': image,
+      };
 }
