@@ -10,7 +10,8 @@ class ChangePassword extends StatefulWidget {
 class _ChangePasswordState extends State<ChangePassword> {
   GlobalKey<FormState> formKey = GlobalKey();
   TextEditingController password = TextEditingController();
-  bool obscureText = true;
+
+  bool obscureText = true, loading = false;
 
   @override
   Widget build(BuildContext context) => AlertDialog(
@@ -21,53 +22,75 @@ class _ChangePasswordState extends State<ChangePassword> {
         content: SingleChildScrollView(
           child: Form(
             key: formKey,
-            child: TextFormField(
-              decoration: InputDecoration(
-                labelText: 'password',
-                prefixIcon: const Icon(Icons.key),
-                suffixIcon: IconButton(
-                  onPressed: () => setState(() => obscureText = !obscureText),
-                  icon: Icon(obscureText ? Icons.remove_red_eye : Icons.lock),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Password',
+                  style: Theme.of(context).textTheme.bodySmall,
                 ),
-              ),
-              controller: password,
-              keyboardType: TextInputType.visiblePassword,
-              obscureText: obscureText,
-              textInputAction: TextInputAction.done,
-              validator: (value) {
-                if (value!.isEmpty) {
-                  return '* required';
-                } else if (value.length < 6) {
-                  return '* must be 6 characters';
-                } else {
-                  return null;
-                }
-              },
-              onFieldSubmitted: (value) => validate(),
+                const SizedBox(height: 8),
+                TextFormField(
+                  style: Theme.of(context)
+                      .textTheme
+                      .bodySmall!
+                      .copyWith(fontWeight: FontWeight.normal),
+                  decoration: InputDecoration(
+                    filled: true,
+                    fillColor: Colors.white,
+                    prefixIcon: const Icon(Icons.key),
+                    suffixIcon: IconButton(
+                      onPressed: () =>
+                          setState(() => obscureText = !obscureText),
+                      icon:
+                          Icon(obscureText ? Icons.remove_red_eye : Icons.lock),
+                    ),
+                  ),
+                  controller: password,
+                  keyboardType: TextInputType.visiblePassword,
+                  obscureText: obscureText,
+                  textInputAction: TextInputAction.done,
+                  validator: (value) {
+                    if (value!.isEmpty) {
+                      return '* required';
+                    } else if (value.length < 6) {
+                      return '* must be 6 characters';
+                    } else {
+                      return null;
+                    }
+                  },
+                  onFieldSubmitted: (value) => validate(),
+                ),
+              ],
             ),
           ),
         ),
 
         // Register
-        actionsAlignment: MainAxisAlignment.center,
         actions: [
           ElevatedButton(
             onPressed: () => validate(),
-            child: const Text('register'),
+            child: loading
+                ? const CircularProgressIndicator()
+                : const Text('register'),
           ),
         ],
       );
 
   void validate() {
+    setState(() => loading = true);
+
     if (formKey.currentState!.validate()) {
       try {
-        FirebaseAuth.instance.currentUser!.updatePassword(password.text);
-
-        Navigator.pop(context);
-        Navigator.pop(context);
-      } on FirebaseAuthException catch (error) {
-        errorSnackBar(error.message.toString());
+        FirebaseAuth.instance.currentUser!
+            .updatePassword(password.text)
+            .then((value) => page(context: context, page: const Main()));
+      } catch (error) {
+        errorSnackBar(context, error.toString());
       }
     }
+
+    setState(() => loading = false);
   }
 }
